training/requirements.txt
albumentations==1.1.0
lpips==0.1.4
numpy==1.21.2
Pillow==8.4.0
torch==1.10.0
torchvision==0.10.0
tqdm==4.62.3
